More formatting

+++ contracts/AMM.sol
@@ -11,7 +11,7 @@ contract AMM is Ownable {
     using LibMath for uint256;
     using LibMath for int256;
 
-    uint256 constant DECIMAL_FACTOR = 10 ** 9;
+    uint256 constant DECIMAL_FACTOR = 10**9;
 
     /* provided for semantic clarity */
     uint256 constant SECONDS_PER_MINUTE = 60;
@@ -34,7 +34,7 @@ contract AMM is Ownable {
         Balances.Position target;
         Balances.Position foobarPosition; /* TODO: what is this for? */
         uint256 liquidityShareSupply;
-        mapping (address => uint256) liquidityShares;
+        mapping(address => uint256) liquidityShares;
     }
 
     Pool highFeePool;
@@ -56,9 +56,7 @@ contract AMM is Ownable {
      * @notice Creates a new Tracer dnAMM instance
      *
      */
-    constructor(
-        address _tracer
-    ) Ownable() {
+    constructor(address _tracer) Ownable() {
         tracer = ITracerPerpetualSwaps(_tracer);
     }
 
@@ -100,29 +98,33 @@ contract AMM is Ownable {
         /* retrieve the fair price for this market */
         int256 fairPrice = getFairPrice(); /* TODO: get fair price */
 
-        pool.foobarPosition.base  = pool.foobarPosition.base + amount;
-        pool.target.quote += amount / getFairPrice() * int(maxLeverage);
-        pool.actual.quote += amount / getFairPrice() * int(maxLeverage);
+        pool.foobarPosition.base = pool.foobarPosition.base + amount;
+        pool.target.quote += (amount / getFairPrice()) * int256(maxLeverage);
+        pool.actual.quote += (amount / getFairPrice()) * int256(maxLeverage);
 
         uint256 liquidityIncrease = 0;
 
-        if(pool.target.base == 0) { /* pool is empty */
+        if (pool.target.base == 0) {
+            /* pool is empty */
             require(amount > 0);
             lastFeeUpdate = block.number;
-            liquidityIncrease = uint(amount);
+            liquidityIncrease = uint256(amount);
             pool.liquidityShares[msg.sender] += liquidityIncrease;
             pool.liquidityShareSupply += liquidityIncrease;
-        } else if(pool.target.base > 0) { /* pool already has liquidity */
-            liquidityIncrease = uint(amount * pool.target.base / pool.actual.base);
+        } else if (pool.target.base > 0) {
+            /* pool already has liquidity */
+            liquidityIncrease = uint256((amount * pool.target.base) / pool.actual.base);
 
-            pool.liquidityShares[msg.sender] += uint(liquidityIncrease *
-                pool.liquidityShareSupply / uint(pool.target.base));
-            pool.liquidityShareSupply += uint(liquidityIncrease *
-                pool.liquidityShareSupply / uint(pool.target.base));
+            pool.liquidityShares[msg.sender] += uint256(
+                (liquidityIncrease * pool.liquidityShareSupply) / uint256(pool.target.base)
+            );
+            pool.liquidityShareSupply += uint256(
+                (liquidityIncrease * pool.liquidityShareSupply) / uint256(pool.target.base)
+            );
         }
 
         /* update pool's target and position (in *base* currency) */
-        pool.target.base += int(liquidityIncrease);
+        pool.target.base += int256(liquidityIncrease);
         pool.actual.base += amount;
     }
 
@@ -141,32 +143,31 @@ contract AMM is Ownable {
             pool = lowFeePool;
         }
 
-        require(int(highFeePool.liquidityShares[msg.sender]) >= amount,
-            "Insufficient LP share balance");
+        require(int256(highFeePool.liquidityShares[msg.sender]) >= amount, "Insufficient LP share balance");
         syncWithPosition();
-        pool.liquidityShares[msg.sender] -= uint(amount);
+        pool.liquidityShares[msg.sender] -= uint256(amount);
 
         /* transfer base currency to msg.sender */
-        pool.foobarPosition.base -= (pool.actual.base * amount) / int(pool.liquidityShareSupply);
+        pool.foobarPosition.base -= (pool.actual.base * amount) / int256(pool.liquidityShareSupply);
         // lpBasePosition += (pool.actual.base * amount) / pool.liquidityShares; /* TODO: API issue */
-        pool.actual.base -= (pool.actual.base * amount) / int(pool.liquidityShareSupply);
+        pool.actual.base -= (pool.actual.base * amount) / int256(pool.liquidityShareSupply);
 
         /* update the BaseTarget in order to remain proportionate */
-        pool.target.base -= ((pool.target.base / pool.actual.base) * amount) /
-            int(pool.liquidityShareSupply);
-        pool.foobarPosition.base -= (pool.target.base * amount) / int(pool.liquidityShareSupply);
+        pool.target.base -= ((pool.target.base / pool.actual.base) * amount) / int256(pool.liquidityShareSupply);
+        pool.foobarPosition.base -= (pool.target.base * amount) / int256(pool.liquidityShareSupply);
 
         /* if the AMM is exposed at the time of withdrawal, the withdrawing LP
          * gets their portion of that exposure */
-        pool.foobarPosition.quote -= ((pool.target.quote - pool.actual.quote) * amount) /
-            int(pool.liquidityShareSupply);
+        pool.foobarPosition.quote -=
+            ((pool.target.quote - pool.actual.quote) * amount) /
+            int256(pool.liquidityShareSupply);
         // lpPositionQuote += ((pool.target.quote - pool.actual.quote) * amount) / /* TODO: API issue */
         //     pool.liquidityShares;
-        pool.actual.quote -= (pool.target.quote - pool.actual.quote) * amount /
-            int(pool.liquidityShareSupply);
-        pool.foobarPosition.quote -= ((pool.target.quote - pool.actual.quote) * amount) /
-            int(pool.liquidityShareSupply);
-        pool.liquidityShareSupply -= uint(amount);
+        pool.actual.quote -= ((pool.target.quote - pool.actual.quote) * amount) / int256(pool.liquidityShareSupply);
+        pool.foobarPosition.quote -=
+            ((pool.target.quote - pool.actual.quote) * amount) /
+            int256(pool.liquidityShareSupply);
+        pool.liquidityShareSupply -= uint256(amount);
     }
 
     /**
@@ -183,20 +184,23 @@ contract AMM is Ownable {
         require(highFeePool.target.base > 0 || 0 < lowFeePool.target.base);
         require(daysSinceFeeUpdate > 1);
 
-        uint256 skew = uint((highFeePool.target.base / (highFeePool.target.base + lowFeePool.target.base)) * int(DECIMAL_FACTOR));
+        uint256 skew = uint256(
+            (highFeePool.target.base / (highFeePool.target.base + lowFeePool.target.base)) * int256(DECIMAL_FACTOR)
+        );
 
-        uint256 adjustedMinutesSinceFeeUpdate = (block.timestamp - lastFeeUpdate -
+        uint256 adjustedMinutesSinceFeeUpdate = (block.timestamp -
+            lastFeeUpdate -
             (SECONDS_PER_MINUTE * MINUTES_PER_HOUR * HOURS_PER_DAY)) / SECONDS_PER_MINUTE;
 
         /* calculate fee */
-        if(skew > 500000000) {
-            lowFee = LibMath.min(250000000 * DECIMAL_FACTOR, lowFee * (4 * (skew - 500000000) ** 2 + 1));
-        } else if(skew < 500000000) {
+        if (skew > 500000000) {
+            lowFee = LibMath.min(250000000 * DECIMAL_FACTOR, lowFee * (4 * (skew - 500000000)**2 + 1));
+        } else if (skew < 500000000) {
             //lowFee = lowFee * (-2 * (skew - 500000000)**2 + 1);
         }
 
         /* handle keeper payment */
-        uint256 keeperPayment = 10000000 * (adjustedMinutesSinceFeeUpdate) ** 3;
+        uint256 keeperPayment = 10000000 * (adjustedMinutesSinceFeeUpdate)**3;
         highFeePool.liquidityShareSupply += keeperPayment / 2;
         lowFeePool.liquidityShareSupply += keeperPayment / 2;
         highFeePool.liquidityShares[msg.sender] += keeperPayment / 2;
@@ -222,39 +226,40 @@ contract AMM is Ownable {
 
         syncWithPosition();
 
-        if(pool.actual.base == pool.target.base) {
+        if (pool.actual.base == pool.target.base) {
             require(amount < 0);
         } else {
             require(pool.actual.base < pool.target.base);
         }
 
-        traderPositionBase -= int(amount);
+        traderPositionBase -= int256(amount);
         uint256 maxLeverage = getMaxLeverage(); // TODO: get max leverage
         int256 fairPrice = getFairPrice(); // TODO: get fair price
-        uint256 feeCollected = uint(amount) * lowFee * 2;
+        uint256 feeCollected = uint256(amount) * lowFee * 2;
 
-        if(pool.actual.base + amount < 0) {
-            feeCollected += (0 - (uint(pool.actual.base + amount))) / maxLeverage;
+        if (pool.actual.base + amount < 0) {
+            feeCollected += (0 - (uint256(pool.actual.base + amount))) / maxLeverage;
         }
 
-        if(amount > 0) {
-            amount -= int(feeCollected);
+        if (amount > 0) {
+            amount -= int256(feeCollected);
         }
 
-        if(amount < 0) {
-            amount += int(feeCollected);
+        if (amount < 0) {
+            amount += int256(feeCollected);
         }
 
-        pool.target.quote += int(
-            int(feeCollected) / getFairPrice() * int(maxLeverage) * pool.target.quote / pool.actual.quote);
-        pool.actual.quote += int(feeCollected) * pool.target.base / pool.actual.base;
+        pool.target.quote += int256(
+            ((int256(feeCollected) / getFairPrice()) * int256(maxLeverage) * pool.target.quote) / pool.actual.quote
+        );
+        pool.actual.quote += (int256(feeCollected) * pool.target.base) / pool.actual.base;
 
         if (amount > 0) {
             require(amount <= pool.target.base - pool.actual.base);
         }
 
-        uint256 exposure = uint(((pool.actual.base + amount / 2) - pool.target.base) / pool.target.base);
-        int256 tradePrice = getFairPrice() + int(sensitivity) * getFairPrice() * int(exposure);
+        uint256 exposure = uint256(((pool.actual.base + amount / 2) - pool.target.base) / pool.target.base);
+        int256 tradePrice = getFairPrice() + int256(sensitivity) * getFairPrice() * int256(exposure);
 
         /* `quotePayment` is the amount of the quote asset LEAVING the AMM.
          * If `amount` is positive, then so is `quotePayment`.
@@ -293,7 +298,7 @@ contract AMM is Ownable {
 
         syncWithPosition();
 
-        if(pool.actual.quote == pool.target.quote) {
+        if (pool.actual.quote == pool.target.quote) {
             require(amount < 0);
         } else {
             require(pool.actual.quote < pool.target.quote);
@@ -301,28 +306,28 @@ contract AMM is Ownable {
 
         traderPositionQuote -= amount;
         uint256 maxLeverage = getMaxLeverage();
-        uint256 exposure = uint((pool.target.quote - (pool.actual.quote - amount / 2)) /
-                                pool.target.quote);
-        int256 highFeePrice = getFairPrice() + int(sensitivity) * getFairPrice() * int(exposure);
-        uint256 feeCollected = uint(amount) * lowFee * 2 * uint(highFeePrice);
+        uint256 exposure = uint256((pool.target.quote - (pool.actual.quote - amount / 2)) / pool.target.quote);
+        int256 highFeePrice = getFairPrice() + int256(sensitivity) * getFairPrice() * int256(exposure);
+        uint256 feeCollected = uint256(amount) * lowFee * 2 * uint256(highFeePrice);
 
         if (pool.actual.quote + amount < 0) {
-            feeCollected += (0 - (uint(pool.actual.quote + amount))) / maxLeverage * uint(getFairPrice());
+            feeCollected += ((0 - (uint256(pool.actual.quote + amount))) / maxLeverage) * uint256(getFairPrice());
         }
 
-        traderPositionBase -= int(feeCollected);
-        pool.foobarPosition.base += int(feeCollected);
-        pool.target.quote += int(feeCollected) / getFairPrice() * int(maxLeverage) *
-                                pool.target.quote / pool.actual.quote;
-        pool.actual.quote += int(feeCollected) / getFairPrice() * int(maxLeverage);
-        pool.target.base += int(feeCollected) * pool.target.base / pool.actual.base;
-        pool.actual.base += int(feeCollected);
+        traderPositionBase -= int256(feeCollected);
+        pool.foobarPosition.base += int256(feeCollected);
+        pool.target.quote +=
+            ((int256(feeCollected) / getFairPrice()) * int256(maxLeverage) * pool.target.quote) /
+            pool.actual.quote;
+        pool.actual.quote += (int256(feeCollected) / getFairPrice()) * int256(maxLeverage);
+        pool.target.base += (int256(feeCollected) * pool.target.base) / pool.actual.base;
+        pool.actual.base += int256(feeCollected);
 
         if (amount > 0) {
             require(amount <= pool.target.base - pool.actual.base);
         }
 
-        int256 tradePrice = getFairPrice() + int(sensitivity) * getFairPrice() * int(exposure);
+        int256 tradePrice = getFairPrice() + int256(sensitivity) * getFairPrice() * int256(exposure);
         int256 basePayment = tradePrice * amount;
         pool.foobarPosition.quote += amount;
         pool.actual.quote += amount;
@@ -332,7 +337,7 @@ contract AMM is Ownable {
     }
 
     /*********************************** PRIVATE ******************************/
-    
+
     function syncPoolWithPosition(Pool storage pool) internal {
         /* targets change proportional to the change in base */
         pool.target.base *= pool.actual.base / pool.foobarPosition.base;
@@ -341,7 +346,7 @@ contract AMM is Ownable {
 
         /* check that at least one currency is within target (note that it
          * should not be possible for this condition to be true) */
-        if(pool.actual.base < pool.target.base && pool.actual.quote < pool.target.quote) {
+        if (pool.actual.base < pool.target.base && pool.actual.quote < pool.target.quote) {
             /* reduce quote target to satisfy safety condition */
             pool.target.quote = pool.actual.quote - (pool.target.base - pool.actual.base) / getFairPrice();
 
@@ -369,4 +374,3 @@ contract AMM is Ownable {
         /* TODO: getFairPrice */
     }
 }
-
